Add realm pod, fix project config, remove bridging headers
Realm pod wasn't compiling. Found that Pods/** custom header path in the targets was causing the issue. After cleaning up pods it was possible to remove this custom header and everything works fine. Also removed the bridging headers as part of possible fix, doesn't seem to be related though but it's better without bridging headers.