Replace core data provider with realm provider and minor related fixes
Core data provider is still in the project because inventory uses it, when everything is ported to realm all core data code will be removed.

Also added loading product and section from remote when adding a list item, if it's not in db. This is not complete considering all possible use cases which can happen using multiple devices and offline/online. But it is an improvement against loading only from local database, as now a second, online, logged in device will also be able to find them (this can happen only if the second device is not synchronized, of course, otherwise the data should also be in its local database).