iOS: Fix crash opening history second time (Realm)
One cause was that the mapper wasn't setting the inventory field in the database history item
since the realm vars have a default value this inserts an empty inventory in the db - which causes a bad access while being loaded, probably because the primary key (uuid) is not set in the empty object?
So error on our side but also Realm, it shouldn't allow to insert invalid objects instead of crashing when loaded

The other cause was that we delete the inventory table on inventory sync -to overwrite with server content- and forgot to call history item sync after it. This has another funny effect in realm - the reference from the history table to the inventory is just deleted (the inventory which is re-added, which has equal content, is ignored). So when we try to access inventory in the loaded history objs, crash!

TODO: database needs proper error handling. Pass result to client using providers, like remote. So this kind of issues don't crash the app