Core/iOS: Fix cart & stash view sometimes not updating correctly
We need to ensure the controllers access basically only the memory cache and database is only "background" processing, otherwise we run into synchronisation issues - when e.g. update item, switch controller automatically which causes a reload and reload is faster than the update we load old data. So because of this controllers should direct interaction of controller should be only with the memory cache, which is synchronous.