iOS: Fix list items to list controller transition looks weird
The problem was that the topbar would shrink due to the constraints with the views below, making price view appear "on top" of the topbar during the transition. Decreasing the priority of the constraint between the views below and the topbar, such that the constraint with the topbar height has a higher priority and doesn't shrink fixed it.