iOS: Replace topbar with normal navbar in manage products. Style topbar like nav bar everywhere else"
Note: Some changes in topbar were done in order to reemplace nav bar with our custom top bar at runtime, such that we preserve both the transition animations of nav bar as well as the possible animations of topbar. This sort of works but needs a bit more work to make the topbar blend perfectly with the navigation bar (such that when they are replaced, user doesn't notice). I noticed it's not really necessary to have "add" button in products screen, and since the other buttons (edit, submit) are not dynamic, there's no reason to use the custom bar. So just replaced it with navigation bar. We keep the changes in topbar though in case this is needed again in the future.