Core/iOS: Implement store products (...)
Now we use store products for store specific information like store and price.
List items reference store products and group/inventory items continue referencing only products, as store information is not relevant to them.

When we add products to list we use the store of the list to fetch/create the corresponding store product.
The functionality is the same for lists with no store, in this case we simply fetch/create store products which have an empty string as store. The "empty string" store is a store like any other.