Core/iOS: Prefill database now with code (...)
The bundled databases turned to be a bad idea since the uuids have to be regenerated - otherwise they will be the same in all clients which causes problems in the server
Regenerating the uuids requires to re-create all the entries, including updating the reference from products to categories so the performance benefit of bundling the db is lost and it's even worser than prefilling programmatically.

Additionally we now use directly db objects to prefill to improve performance.