Core: Change the type of timestamps to Int64 (...)
- Fixes: Timestamps were not correct (we were interpreting the milliseconds sent by server as seconds)
- To simplify things we don't use NSDate anymore but numeric timestamp this way we don't have to worry about timezones. The client doesn't do any operations with the timestamp except sending it back to the server so this should be ok.
= We could have used Double, but this fails when sending the timestamps to the server which expects long, we of course could convert to Int64 before sending to server but this is easy to forget as the parameter dictionaries are untyped.
On the other side we now have to convert the incoming data from the server from Double to Int64 (parsing directly as Int64 is not possible - the cast returns nil). But this is better since in our parser classes we are forced to use the correct types and can't just forget to cast.
- We could have used UInt64, which is better for timestamps but it's not supported by Realm.