Core: Fix can't add list item with a section that previously disappeared from list as result of deleting all the list (...)
This triggered a server error, since we would create a new section, which causes a unique violation in the server as it has same name and list uuid as the existing one.

In detail - when we "empty" sections either by deleting list items inside it or moving them to other sections, the sections is not removed from the database
And when adding a list item with a new section, we were reading only the list items in the memory provider in order to retrieve the section. The not referenced section of course, wasn't found here. So a new section was crated.

The reason we don't delete section after emptying is a combination of UX and consistency, see comments made in code in this commit.

So now we check first for the section in the database and reference it if it exists.

The server will be updated anyway to solve the contraint duplication by appending (number) to the name instead of returning an error.