Core: Re-fix deleted (synced) list item offline appears again after sync (...)
Turns the fix in 3780a21 was not a good idea because the timetamp in the model objects is not always up to date! When we do the bg update the ui objects are not necessarily updated, so when we delete and write the timestamp of the model object in the tombstone it may be that this timestamp is old or 0.
So now we load the db object in order to get the timestamp.