Add dirty parameter to inventories and groups and don't add tombstones on websocket (...)
This replicated previous commits in which this was done for lists
I should probably have used default parameter = true for dirty, for most methods this is only false when it comes from websockets
It turned to be necessary to change a lot more places than initially expected. At the end let it like this because revert looked a bit finicky and don't have time.