Core/iOS: Don't use models as dictionary key (...)
hashValue in models was inconsistent with equals. For equals we include all fields including the server timestamp. For dictionaries we prefer to ensure we work only with the semantic identity of the object (uuid) to avoid possible inconsistencies in possible (un)expected situations, like different server timestamps or generally different content, which for some reason hasn't synced yet.

For example, when we receive list item via websocket this may have updated the section for the sender and in the server, but the timestamps of sections in the table view of the receiver don't have these timestamps yet so when we try to ininsert the listitem in the table view, with current code that used dictionary with sections as keys it would insert a new section instead of using the existing one.

So models don't conform to hashable anymore and we use uuids directly as keys.