Core/iOS: On login with a new device, do a full download instead of sync (...)
If we do sync we will upload all the prefilled products of the new device and since the uuids are different but the same semantic uniques, we get all the products repeated (with a (1)) at the end.
To prevent this we now check if it's a new and not-first device (user has not registered with this device and login is successful), if yes we do the full download (local db overwrite).
To avoid surprises we show the user a popup offering the full download, if user cancels there's no download and the login is cancelled.