iOS: Use section instead of category color in products quick add when in list
The reason for this is that when we add the item to the list, it's added using the section color and when the section color is different than category color, it looks like a bug to the user - it shows in one color in the quick add and in another in the list. So now we override the category color with the section color if existent in the quick add.