Core: Workaround for empty storeproducts/products/categories (...)
Sometimes the on sync client sends these empty objects. This causes a sync error since server has validation rules e.g. min length for uuid. The reason of these objects, should be the fallback of list item, when its store product was removed without removing the list item. Searched in code for such situation but coulnd't find it. When producs etc. are removed, the dependencies are also removed. Clearly missing something.

For now we filter such empty objects of the sync request - since we overwrite everything on response, the objects are gone after sync.