Fix in cart note not completely at the right side
After layoutIfNeeded() in setMode the width of the +/- buttons is not yet updated despite we just set the constraint to 0. So we now use directly the constraint constant in intrinsic size.

This doesn't happen in todo list because intrinsic size is called 2x - the second time it returns the updated width. This update seems to be triggered during the controller's expand animation.

Probably it's better to force update constraints (?) but this works.