Disable swipe to delete in list item cells
The reason is that it looks weird here as we have reorder control and it moves only the (custom) content view which the delete button between this view and the reorder control instead of at the right side. It's particularly weird as it's still possible to delete using default iOS animation which shows it correctly at the right side. In inventory items we can do swipe to delete because it doesn't have reorder control.