Quick and dirty fix for list item not saving note when added via contact form Very ugly fix, had to pass note parameter to everything (inventory, ingredients etc)
The cause of the bug was that adding new items was split in 2 steps in order to first add the item first to the collection view and then animate towards the items table view. The first step (onSubmitAddEditItem2) - where we pass the note in ListItemInput - adds only the store product and the second (showAddedItem) actually creates the list item and doesn't have a note parameter (this is meant to used by other types of items that don't have a note). The fix was to just add an optional note parameter to showAddedItem which is only used in list items case.

Todo: This implementation should be improved.